+++ CHANGELOG.md
@@ -21,6 +21,7 @@
     * Optimized `Flex Pages` collection filtering
 1. [](#bugfix)
     * Regression: Fixed Grav update bug [#2722](https://github.com/getgrav/grav/issues/2722)
+    * Grav 1.7: Reverted `$object->getStorageKey()` interface as it was not a good idea, added `getMasterKey()` for pages
 
 # v1.7.0-rc.1
 ## 11/06/2019